Move `r_stop_internal()` implementation to rlang
