+++ src/Makevars
@@ -42,6 +42,7 @@ internal-files = \
         internal/ast-rotate.c \
         internal/attr.c \
         internal/call.c \
+        internal/cnd.c \
         internal/dots.c \
         internal/dots-ellipsis.c \
         internal/encoding.c \
