+++ src/internal/decl/cnd-decl.h
