+++ src/internal/init.c
@@ -246,6 +246,7 @@ void R_init_rlang(DllInfo* dll) {
   R_RegisterCCallable("rlang", "rlang_quo_get_expr",        (DL_FUNC) &ffi_quo_get_expr);
   R_RegisterCCallable("rlang", "rlang_quo_set_env",         (DL_FUNC) &ffi_quo_set_env);
   R_RegisterCCallable("rlang", "rlang_quo_set_expr",        (DL_FUNC) &ffi_quo_set_expr);
+  R_RegisterCCallable("rlang", "rlang_stop_internal",       (DL_FUNC) &rlang_stop_internal);
 
   r_obj* r_as_function(r_obj* x, const char* arg);
   R_RegisterCCallable("rlang", "rlang_as_function",         (DL_FUNC) &r_as_function);
